+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 ## Unreleased
+## 1.153.0 (December, 08 2023)
+ENHANCEMENTS:
+* resource/spotinst_elastigroup_aws: Added `logging` object support.
+
+## 1.152.0 (December, 07 2023)
+NOTES:
+* Migration of all unit tests to run on new test account.
 
 ## 1.151.1 (December, 06 2023)
 NOTES: